FIX : Yahoo background bug
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@13839 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ QMapControl/browserimagemanager.cpp
@@ -71,10 +71,10 @@ void BrowserWebPage::javaScriptAlert ( QWebFrame * frame, const QString & msg )
 		Q_ASSERT(tokens.size() == 5);
 
 		int ox, oy, x1, y1;
-        ox = tokens[1].toInt();
-        oy = tokens[2].toInt();
-        x1 = tokens[3].toInt();
-        y1 = tokens[4].toInt();
+		ox = int(tokens[1].toDouble());
+		oy = int(tokens[2].toDouble());
+		x1 = int(tokens[3].toDouble());
+		y1 = int(tokens[4].toDouble());
 
 		qDebug() << ox << ", " << oy << ", " << x1 << "," << y1;
